fix #71, support for redirecting static symbols
This patch adds experimental support for redirecting static symbols. It
does this by patching the first instruction of the symbol to be a
jmp, and then adding a new RELOC_REL32 relocation pointing at the
operand.

In theory this means we can now re-implement functions that don't work
or need to be patched in c.