for #57, add a -e command line option to eval a macro.
This commit adds a `-e` command line option to evaluate a macro
as soon as 123 is ready to start processing input. Once the
macro has played, you can continue to interact with 123 normally.

For example, to add a new sheet on startup:

```
$ 123 -e '/wisa~'
```

The intention is to use this as part of some automatic testing.