fix #57, add some tests
This commit adds the expected (or "gold") output from opening
most of the sample worksheets that come with 1-2-3. There is a
script that checks the output is as expected.

The intention is that this is the foundation of being able to
make bigger changes without having to worry about breaking old
spreadsheets.