for #73, translate UTF-8 sequences to LMBCS on input.
This is the first step in improving i18n support. If any
UTF-8 sequences have LMBCS encodings, translate them on
input.

These characters are stored as LMBCS internally, and you
can differentiate them with @CODE, but they are not
displayed correctly (they are transliterated to ASCII, see
the 1-2-3R3.1 manual, Appendix 2).

The next part of this change will be displaying them
as UTF-8.