add complete showme implementation for #94
It turns out we can't short circuit the showme logic, because the full
code is needed for the /Data Parse Format editor. This is a more complete
reimplementation of the original.

It's unfortunate so much work is needed for such a minor feature, but
I don't want to break anything.